refactor: home의 shared-compose 이동
