+++ presentation/src/main/kotlin/team/duckie/app/android/presentation/IntroActivity.kt
@@ -24,6 +24,7 @@ import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.first
 import team.duckie.app.android.feature.datastore.PreferenceKey
 import team.duckie.app.android.feature.datastore.dataStore
+import team.duckie.app.android.feature.ui.home.screen.HomeActivity
 import team.duckie.app.android.feature.ui.onboard.OnboardActivity
 import team.duckie.app.android.presentation.screen.IntroScreen
 import team.duckie.app.android.util.compose.ToastWrapper
@@ -55,7 +56,7 @@ class IntroActivity : BaseActivity() {
 
             when (isOnboardFinished) {
                 true -> toast("온보딩 끝냄")
-                false -> changeActivityWithAnimation<OnboardActivity>()
+                false -> changeActivityWithAnimation<HomeActivity>()
                 else -> Unit // null
             }
         }
